Рефакторит код, улучшает структуру проекта
