--- src/components/page-main/trip-days/trip-days.js
@@ -1,37 +1,0 @@
-import {TimeInMs} from '../../../helpers/constants.js';
-
-const getDaysDifference = (tripDays) => {
-  const tripDaysCounters = [];
-  const startCounterNumber = 0;
-
-  for (let i = tripDays.length - 1; i > 0; i--) {
-    const dayDifference = (tripDays[i] - tripDays[i - 1]) / TimeInMs.DAY;
-    tripDaysCounters.push(dayDifference);
-  }
-
-  tripDaysCounters.reverse();
-  tripDaysCounters.unshift(startCounterNumber);
-
-  return tripDaysCounters;
-};
-
-const getTripDays = (tripDaysCounters) => {
-  const tripDays = Array.of(1);
-
-  for (let i = 1; i < tripDaysCounters.length; i++) {
-    tripDays.push(tripDays[i - 1] + tripDaysCounters[i]);
-  }
-
-  return tripDays;
-};
-
-const getTripDaysWithDates = (tripDays, uniqueTripEventsDates) => {
-  return tripDays.map((day, index) => {
-    return {
-      counter: day,
-      date: uniqueTripEventsDates[index],
-    };
-  });
-};
-
-export {getDaysDifference, getTripDays, getTripDaysWithDates};
