+++ src/components/page-main/trip-events/get-event-time-format.js
@@ -1,0 +1,9 @@
+const getEventTimeFormat = (time) => {
+  const timeValues = Array.of(time.getHours(), time.getMinutes()).map((value) => {
+    return value < 10 ? `0` + value : value;
+  });
+
+  return timeValues.join(`:`);
+};
+
+export {getEventTimeFormat};
