+++ src/components/page-main/trip-events/get-time-difference.js
@@ -15,12 +15,4 @@ const getTimeDifference = (start, end) => {
   return daysString + hoursString + minutesString;
 };
 
-const getEventTimeFormat = (time) => {
-  const timeValues = Array.of(time.getHours(), time.getMinutes()).map((value) => {
-    return value < 10 ? `0` + value : value;
-  });
-
-  return timeValues.join(`:`);
-};
-
-export {getTimeDifference, getEventTimeFormat};
+export {getTimeDifference};
