+++ src/mocks/generate-trip-events.js
@@ -4,7 +4,7 @@ import {getRandomDescription} from './get-random-description.js';
 import {getRandomOffers} from './get-random-offers.js';
 import {generateStartDate, generateEndDate} from './generate-dates.js';
 import {eventTypes, eventActionsMap, eventDestinations, eventOffers} from './trip-event-mocks.js';
-import {getTimeDifference} from '../components/page-main/trip-events/trip-event-time.js';
+import {getTimeDifference} from '../components/page-main/trip-events/get-time-difference.js';
 
 const generateTripEvent = () => {
   const hasOffers = Math.random() > 0.5;
